data type change to track upload progress s3 aws

+++ app/src/main/java/com/lifespandh/ireflexions/utils/network/UploadFileStatus.kt
@@ -1,7 +1,7 @@
 package com.lifespandh.ireflexions.utils.network
 
 sealed class UploadFileStatus{
-    data class FileStatus(val status: Int): UploadFileStatus()
+    data class FileStatus(val status: Double): UploadFileStatus()
     data class Error(val exception: Throwable): UploadFileStatus()
     data class Complete(val s3Url: String): UploadFileStatus()
     data class Start(val start: Boolean): UploadFileStatus()
